update isDone and userFixture

+++ src/Controller/TaskController.php
@@ -3,9 +3,11 @@
 namespace App\Controller;
 
 use App\Entity\Task;
+use App\Entity\User;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use phpDocumentor\Reflection\Types\Boolean;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -21,7 +23,6 @@ class TaskController extends AbstractController
         return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findAll()]);
     }
 
-
     #[Route('/tasks/create', name: 'task_create')]
     public function create(Request $request, EntityManagerInterface $em)
     {
@@ -33,7 +34,8 @@ class TaskController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
 
             $task->setCreatedAt(new \DateTime());
-            $task->toggle(false);
+            $task->setOwner($this->getUser());
+            $task->setIsDone(false);
             $em->persist($task);
             $em->flush();
 
@@ -71,7 +73,7 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
     public function toggleTask(Task $task,EntityManagerInterface $em)
     {
-        $task->toggle(!$task->isDone());
+        $task->setIsDone(!$task->isDone());
         $em->persist($task);
         $em->flush();
 
@@ -83,11 +85,32 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/delete', name: 'task_delete')]
     public function deleteTask(Task $task,EntityManagerInterface $em)
     {
+        /** @var User $user */
+        $user = $this->getUser();
+
+        if(!$this->testAsRightToDelete($task,$user)){
+            $this->addFlash('error', 'Vous devez être propiétaire de la tâche pour la supprimer');
+            return $this->redirectToRoute('task_list');
+        }
+
         $em->remove($task);
         $em->flush();
 
         $this->addFlash('success', 'La tâche a bien été supprimée.');
 
         return $this->redirectToRoute('task_list');
+    }
+
+    private function testAsRightToDelete(Task $task,User $user) : bool
+    {
+        if(!$task->getOwner() && !in_array('ROLE_ADMIN', $user->getRoles())){
+            return false;
+        }
+
+        if($task->getOwner() && $task->getOwner() != $user){
+            return false;
+        }
+
+        return true;
     }
 }
